moved types and validation out of localstorage file
