+++ src/app/_utils/localStorageTypes.ts
@@ -1,5 +1,4 @@
 import { Brand } from './typehelpers'
-import { GlobalStateProperties } from './zustandLocalStorage'
 
 //----------------------------------------------
 //------------Local Storage Values ------------------
@@ -111,4 +110,87 @@ export const lsDrops = 'drops'
 export const lsBlowups = 'blowups'
 export const lsMisc = 'misc'
 
-export type LocalStorageKeys = keyof GlobalStateProperties
+export type LocalStorageKeys = keyof GlobalStateProperties/**
+ * Types of Properties on the Zustand Local Storage Global
+ */
+
+export type GlobalStateProperties = {
+  [lsFlows]: Flow[];
+  [lsUserMoves]: {
+    [lsToprock]: string[];
+    [lsFootwork]: string[];
+    [lsPower]: string[];
+    [lsFreezes]: string[];
+    [lsFloorwork]: string[];
+    [lsSuicides]: string[];
+    [lsDrops]: string[];
+    [lsBlowups]: string[];
+    [lsMisc]: string[]
+  };
+  [lsUserLearning]: Move[];
+  [lsDanceList]: string[]
+}
+/**
+ * Types for the Zustand Middlewares, used with StateCreator generic
+  For some reason when used makes TS errors. Possibly a package issue rather than dev error.
+ */
+type ZustandMiddlewareMutators = [
+  ['zustand/persist', ZustandGlobalStore],
+  ['zustand/immer', never]
+]
+export type GlobalStatePropertiesV0 = {
+  [lsFlows]: Flow[];
+  [lsUserMoves]: string[];
+  [lsUserLearning]: Move[];
+  [lsDanceList]: string[]
+}
+/**
+ * Type that has all properties and methods of globalstate
+ */
+
+export type ZustandGlobalStore = GlobalStateProperties & {
+  //============root level===============
+  //-----Setters (Root Level Keys)-----
+  setLsFlows: (flows: Flow[]) => void
+  setLsUserMoves: (moves: GlobalStateProperties[typeof lsUserMoves]) => void
+  setLsUserLearning: (learning: Move[]) => void
+  setDanceList: (list: string[]) => void
+  //-----Getters (Root level keys )------
+  getLsFlows: () => Flow[]
+  getLsUserMoves: () => GlobalStateProperties[typeof lsUserMoves]
+  getLsUserLearning: () => Move[]
+  getDanceList: () => string[]
+
+  //============nested================
+  //-------User Move Keys --------
+  setLsUserMovesByKey: (
+    key: keyof GlobalStateProperties[typeof lsUserMoves],
+    values: string[]
+  ) => void
+  getLsUserMovesByKey: (
+    key: keyof GlobalStateProperties[typeof lsUserMoves]
+  ) => string[]
+  //=================================
+  //------Reinitialization----------
+  replaceGlobalState: (state: {
+    state: GlobalStateProperties
+    version: number
+  }) => void
+  resetGlobalState: () => void
+}
+/**
+ * validator if state is version 0 of globalstateproperties
+ * @param state localstorage
+ * @param version state version
+ * @returns boolean
+ */
+export const isGlobalStateV0 = (
+  state: unknown,
+  version: number
+): state is GlobalStatePropertiesV0 => {
+  //wow this is a lazy validation. it'd be nice if zod was used in future to parse
+  if (version === 0) {
+    return true
+  } else return false
+}
+
